Remove too-specific inventory settings.
We no longer want to use the too-old system python interpreter for
everything. Instead, we'll use whatever's on the path.

When running with the scripts, this should always pick up the latest
version.

There's probably no reason to force the shell to be the
system's zsh. If a newer version is installed, it's probably fine to
use that instead.