Turn off the annoying markdown numbering warnings.
The whole point of avoiding explicit numbers is that lists, when
reordered or otherwise modified, do not need to be manually
renumbered. That's the sort of thing computers are good at doing, and
markdown renderers are more than capable of incrementing.

This warning is annoying and counterproductive. And now it's off.